Add tax and mot statuses to Vehicle Details screen also
