+++ feature/dvla/src/main/kotlin/uk/gov/govuk/dvla/navigation/DvlaNavigation.kt
@@ -83,6 +83,7 @@ fun NavGraphBuilder.dvlaGraph(
             )
         ) {
             VehicleDetailsRoute(
+                launchBrowser = launchBrowser,
                 onBack = onBack,
                 modifier = modifier
             )
